create token service and move related functions
